Termux NGROK

+++ termux-ngrok.sh
@@ -1,22 +1,22 @@
-#!/bin/dash
-# Desenvolvido por Tchelo Noob
-atualizar='\033[1;32m'
-echo "$atualizar Atualizando repositório..."
-apt update && upgrade -y
-apt install -y wget
+#!/bin/bash
+# By Marcel0Sousa in github
+# Edit By ELi-Nunez in github
+update='\033[1;32m'
+echo -e "$update Updating Repositories..."
+apt update > /dev/null 2>&1 && upgrade -y > /dev/null 2>&1
+apt install -y wget > /dev/null 2>&1
 clear
 k='\033[01;32m'
-echo
-echo "${k}████████╗ ██████╗██╗  ██╗███████╗██╗      ██████╗ ";
-echo "${k}╚══██╔══╝██╔════╝██║  ██║██╔════╝██║     ██╔═══██╗";
-echo "${k}   ██║   ██║     ███████║█████╗  ██║     ██║   ██║";
-echo "${k}   ██║   ██║     ██╔══██║██╔══╝  ██║     ██║   ██║";
-echo "${k}   ██║   ╚██████╗██║  ██║███████╗███████╗╚██████╔╝";
-echo "${k}   ╚═╝    ╚═════╝╚═╝  ╚═╝╚══════╝╚══════╝ ╚═════╝ ";
-echo "${k} /_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/";
-echo
-echo "Deseja instalar o Ngrok? [Y/n]"
-read opcao
+echo -e
+echo -e "${k}████████╗ ██████╗██╗  ██╗███████╗██╗      ██████╗ ";
+echo -e "${k}╚══██╔══╝██╔════╝██║  ██║██╔════╝██║     ██╔═══██╗";
+echo -e "${k}   ██║   ██║     ███████║█████╗  ██║     ██║   ██║";
+echo -e "${k}   ██║   ██║     ██╔══██║██╔══╝  ██║     ██║   ██║";
+echo -e "${k}   ██║   ╚██████╗██║  ██║███████╗███████╗╚██████╔╝";
+echo -e "${k}   ╚═╝    ╚═════╝╚═╝  ╚═╝╚══════╝╚══════╝ ╚═════╝ ";
+echo -e "${k} /_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/_/";
+echo -e
+read -p "Install Ngrok? [Y/n]" opcao
 case $opcao in
 y)
 echo
@@ -35,28 +35,33 @@ i*86)
 x86_64)
     architectureURL="amd64" ;;
 *)
-    echo "Arquitetura desconhecida"
+    echo "Unknown Architecture"
 esac
 
-wget "https://github.com/tchelospy/NgrokTest/blob/master/ngrok-stable-linux-${architectureURL}.zip?raw=true" -O ngrok.zip
+wget "https://github.com/tchelospy/NgrokTest/blob/master/ngrok-stable-linux-${architectureURL}.zip?raw=true" -O ngrok.zip > /dev/null 2>&1
 unzip ngrok.zip
-cat ngrok > /data/data/com.termux/files/usr/bin/ngrok
+mv ngrok /data/data/com.termux/files/usr/bin
 chmod 700 /data/data/com.termux/files/usr/bin/ngrok
-rm ngrok ngrok.zip
+read -p "Remove ngrok.zip [Y/n]? remove
+if [ $remove = "y" ]
+    rm ngrok ngrok.zip
+then
+    echo "Ok."
+fi
 clear
-echo "${k}███╗   ██╗ ██████╗ ██████╗  ██████╗ ██╗  ██╗";
-echo "${k}████╗  ██║██╔════╝ ██╔══██╗██╔═══██╗██║ ██╔╝";
-echo "${k}██╔██╗ ██║██║  ███╗██████╔╝██║   ██║█████╔╝ ";
-echo "${k}██║╚██╗██║██║   ██║██╔══██╗██║   ██║██╔═██╗ ";
-echo "${k}██║ ╚████║╚██████╔╝██║  ██║╚██████╔╝██║  ██╗";
-echo "${k}╚═╝  ╚═══╝ ╚═════╝ ╚═╝  ╚═╝ ╚═════╝ ╚═╝  ╚═╝";
-echo
-echo "Exemplo de comando (ngrok http 80\nOu ngrok para ajuda)"
+echo -e "${k}███╗   ██╗ ██████╗ ██████╗  ██████╗ ██╗  ██╗";
+echo -e "${k}████╗  ██║██╔════╝ ██╔══██╗██╔═══██╗██║ ██╔╝";
+echo -e "${k}██╔██╗ ██║██║  ███╗██████╔╝██║   ██║█████╔╝ ";
+echo -e "${k}██║╚██╗██║██║   ██║██╔══██╗██║   ██║██╔═██╗ ";
+echo -e "${k}██║ ╚████║╚██████╔╝██║  ██║╚██████╔╝██║  ██╗";
+echo -e "${k}╚═╝  ╚═══╝ ╚═════╝ ╚═╝  ╚═╝ ╚═════╝ ╚═╝  ╚═╝";
+echo -e
+echo -e "Command example (ngrok http 80\nOr ngrok for help)"
 ;;
 
 n)
 clear
-echo "Ngrok não instalado :("
+echo "Instalaton Failed :("
 echo
 esac
